requirements.txt
gradio>=3.0.0
langgraph
groq>=0.37.1
langchain-groq>=1.1.2
python-dotenv>=1.2.2
openai-whisper>=20250625
speechrecognition>=3.16.0
pydantic>=2.0.0

pandas
ipykernel
transformers[torch]
datasets
accelerate
packaging
gtts
faker
pyprojroot
uvicorn
python-multipart
fastapi
